Refactor Go package structure and naming conventions
The main.go file has been not just rewritten but relocated to cmd/go-search to follow Go project layout conventions. Alongside this, methods and variables were renamed using camel case to match the Go convention. This improves code readability and overall adherence to standard Go project organization.
